requirements.txt
streamlit==1.32.0
web3==6.15.1
openai==1.14.0
python-dotenv==1.0.1
requests==2.31.0
pandas==2.2.1
